_build/requirements.txt
# theGuide Python Dependencies
# Install with: pip install -r requirements.txt

# Required
mysql-connector-python>=8.0.0
pika>=1.3.0
requests>=2.28.0
networkx>=3.0
matplotlib>=3.6.0
orjson>=3.8.0

# Windows Service Support (t16o_exchange.guide.*)
pywin32>=306

# Optional (for advanced analysis)
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
